endpoints in users router
