+++ fractal_server/app/routes/auth/users.py
@@ -23,7 +23,10 @@ from ._aux_auth import _get_single_user_with_group_ids
 from fractal_server.app.models import LinkUserGroup
 from fractal_server.app.models import UserGroup
 from fractal_server.app.models import UserOAuth
+from fractal_server.app.models import UserSettings
 from fractal_server.app.routes.auth._aux_auth import _user_or_404
+from fractal_server.app.schemas import UserSettingsRead
+from fractal_server.app.schemas import UserSettingsUpdate
 from fractal_server.app.security import get_user_manager
 from fractal_server.app.security import UserManager
 from fractal_server.logger import set_logger
@@ -196,3 +199,39 @@ async def list_users(
         )
 
     return user_list
+
+
+@router_users.get(
+    "/users/{user_id}/settings/", response_model=UserSettingsRead
+)
+async def get_user_settings(
+    user_id: int,
+    superuser: UserOAuth = Depends(current_active_superuser),
+    db: AsyncSession = Depends(get_async_db),
+) -> UserSettingsRead:
+
+    user = await _user_or_404(user_id=user_id, db=db)
+    user_settings = await db.get(UserSettings, user.user_settings_id)
+    return user_settings
+
+
+@router_users.patch(
+    "/users/{user_id}/settings/", response_model=UserSettingsRead
+)
+async def patch_user_settings(
+    user_id: int,
+    settings_update: UserSettingsUpdate,
+    superuser: UserOAuth = Depends(current_active_superuser),
+    db: AsyncSession = Depends(get_async_db),
+) -> UserSettingsRead:
+    user = await _user_or_404(user_id=user_id, db=db)
+    user_settings = await db.get(UserSettings, user.user_settings_id)
+
+    for k, v in settings_update.dict(exclude_unset=True).items():
+        setattr(user_settings, k, v)
+
+    db.add(user_settings)
+    await db.commit()
+    await db.refresh(user_settings)
+
+    return user_settings
